refactor(portability): Refactored cache_flush_range signature
cache_flush_range signature has been altered to make it portable.

+++ src/arch/armv8/cache.c
@@ -108,14 +108,14 @@ void cache_arch_enumerate(cache_t *dscrp)
     }
 }
 
-void cache_flush_range(void* base, uint64_t size)
+void cache_flush_range(void* base, size_t size)
 {
     uint64_t cache_addr = (uint64_t)base;
     uint64_t ctr = MRS(CTR_EL0);
     uint64_t min_line_size = 1UL << bit_extract(ctr, CTR_DMINLINE_OFF, 
         CTR_DMINLINE_LEN);
 
-    while(cache_addr < ((uint64_t)base + size)){
+    while(cache_addr < ((uint64_t)base + (uint64_t)size)){
         asm volatile (
             "dc civac, %0\n\t" 
             :: "r"(cache_addr));
